projet Hopital Fes/requirements.txt
Django>=4.2.24
djangorestframework
django-cors-headers
